lnd: delete mock.go, move mockSpendNotifier to lntest/mock
